+++ code/crates/test/tests/it/main.rs
@@ -188,7 +188,7 @@ impl TestRunner {
                 enabled: true,
                 // Current test app does not support proposal-only value payload properly as Init does not include valid_round
                 value_payload: ValuePayload::ProposalAndParts,
-                queue_capacity: 100, // Deprecated, derived from `sync.parallel_requests`
+                queue_capacity: 100,
                 p2p: P2pConfig {
                     protocol,
                     discovery: DiscoveryConfig::default(),